refactor: move public classes to tachyon-api
