+++ tachyon-api/src/main/java/dev/tachyonmcp/server/domain/InvalidArgumentException.java
