+++ tachyon-api/src/test/java/dev/tachyonmcp/server/domain/AnnotationsTest.java
